pyproject.toml
[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[project]
name    = "lxd-machine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "blake3",
    "docling",
    "docling-core",
    "fastmcp",
    "lancedb",
    "networkx",
    "ollama",
    "pyahocorasick",
    "pydantic",
    "pyyaml",
    "rich",
    "semchunk",
    "structlog",
    "tree-sitter",
    "tree-sitter-c",
    "tree-sitter-java-orchard",
    "tree-sitter-javascript",
    "tree-sitter-python",
    "tree-sitter-typescript",
    "typer",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lxd"]

[tool.ruff]
line-length    = 100
target-version = "py311"

[tool.ruff.lint]
select  = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore  = [
    "E501",
    "TC001",  # pydantic resolves types at runtime — moving to TYPE_CHECKING breaks it
    "TC002",
    "TC003",
]

[tool.pyright]
pythonVersion                = "3.11"
typeCheckingMode             = "strict"
include                      = ["src"]
reportMissingImports         = true
reportMissingTypeStubs       = false  # ahocorasick, networkx, blake3, lancedb have no stubs
reportUnknownVariableType    = false  # untyped C extensions (ahocorasick, networkx)
reportUnknownMemberType      = false  # untyped C extension method calls
reportUnknownArgumentType    = false  # passing untyped objects to typed functions
reportUnknownParameterType   = false  # parameters typed by untyped libraries
reportMissingTypeArgument    = false  # networkx DiGraph[NodeType, EdgeType] — not required here

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
